Work with db, adding some methods

+++ dolgobot/src/com/redeyesgang/DB/IGetInfo.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface IGetInfo {
-    Map<Integer,Integer> getTotal(long userID) throws SQLException;
+    Map<Long,Long> getTotal(long userID) throws SQLException;
     List<String> getGroupInfo(String groupName);
